emo-nextjs/backend/requirements.txt
# EMO Backend - FastAPI + LangGraph
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# LangChain/LangGraph
langchain-google-genai>=2.0.0,<2.1.0
langchain-core>=0.3.0,<0.4.0
langgraph>=0.2.0,<1.0.0
google-generativeai>=0.8.0,<1.0.0

# Memory
chromadb>=0.5.0

# Google APIs
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
youtube_transcript_api>=0.6.0

# Document parsing
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
pypdf>=4.0.0
